ft_transcendence/requirements.txt
asgiref==3.8.1
attrs==23.2.0
bcrypt==4.1.3
capstone==5.0.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colored-traceback==0.3.0
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
Django==4.2.15
exceptiongroup==1.2.1
flake8==7.1.1
fonttools==4.53.1
h11==0.14.0
idna==3.7
importlib_resources==6.4.0
intervaltree==3.1.0
kiwisolver==1.4.5
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.9.2
mccabe==0.7.0
numpy==2.0.1
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
paramiko==3.4.0
pillow==10.4.0
plumbum==1.8.3
psutil==5.9.8
pwntools==4.12.0
pycodestyle==2.12.1
pycparser==2.22
pycryptodomex==3.20.0
pyelftools==0.31
pyflakes==3.2.0
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==3.1.2
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
ROPGadget==7.4
rpyc==6.0.0
seaborn==0.13.2
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
tqdm==4.66.5
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.1
tzdata==2024.1
unicorn==2.0.1.post1
unix-ar==0.2.1
urllib3==2.2.1
wsproto==1.2.0
zipp==3.20.0
zstandard==0.22.0
django==4.2.15
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.6.0
pyotp==2.9.0